Extracting the toolbar button related methods from EmsMiddlewareController into MiddlewareOperationsMixin.
The button method is now clever and finds out that we are trying to run
the operation for the associated entity (if the @display is set) and
checks that the propriate controller has the operation whitelisted and
only then it runs it. If not it delegates the call to the
ems_common.button method.

Also 1 small change was made into ems_common.rb to allow to tag the
entites when user is on the /{parent_entity}/{id}?display={child_entity}
type of URL.
