+++ app/controllers/ems_middleware_controller.rb
@@ -1,6 +1,7 @@
 class EmsMiddlewareController < ApplicationController
   include EmsCommon
   include Mixins::EmsCommonAngular
+  include MiddlewareOperationsMixin
 
   before_action :check_privileges
   before_action :get_session_data
