+++ app/controllers/middleware_datasource_controller.rb
@@ -17,15 +17,8 @@ class MiddlewareDatasourceController < ApplicationController
     }
   }.freeze
 
-  def button
-    selected_operation = params[:pressed].to_sym
-    if OPERATIONS.key?(selected_operation)
-      selected_ds = identify_selected_entities
-      run_operation(OPERATIONS.fetch(selected_operation), selected_ds)
-      javascript_flash
-    else
-      super
-    end
+  def self.operations
+    OPERATIONS
   end
 
   menu_section :mdl
